requirements.txt

pandas==1.3.4
pillow == 8.4.0

